requirements.txt
crawl4ai
streamlit
sentence_transformers
faiss-cpu